ShardDataTransfer in a separate file
